Remove entry point from soldier object...
Apparently having an entry point on soldiers allows bots to steal human player's bodies while they are out using the RC tank. it's rare, but I've had seen it happen. The result is the human player and a bot occupying the same soldier object. Both players appear to have control over the soldier (not sure if bot could still switch weapons or not). While interesting, the game will crash if a soldier occupied by two players dies. I've not seen bots posses the soldier objects of other buts. It's safe to assume it can happen and may also result in game crashes when said soldier dies.